ai-service/requirements.txt
fastapi
uvicorn
elasticsearch==8.13.2
scikit-learn
numpy
pandas
apscheduler
requests